* V5-7-patches:
  perl/OID/OID.xs: Fix a compiler warning
  configure: Document that --with-perl-modules=verbose is helpful while debugging Makefile.PL files
  testing/fulltests/unit-tests/T003copy_nword_clib: Fix Darwin compiler warnings
  Add support for Darwin 15, 16 and 17
  MIB-II: Restore the declaration of TCP_Count_Connections()
  Windows, MIB-II, system: Read Windows version from registry
  Cygwin, swrun: Avoid using GetVersionEx()
  Win32, MSVC batch build: Embed OpenSSL library path in Makefiles instead of appending it to %LIB%
  Win32: Enable cross-compilation in build.pl
  Windows: Avoid casting a pointer to an integer type with a different size

+++ agent/mibgroup/mibII/tcpTable.c
@@ -371,7 +371,7 @@ tcpTable_next_entry( void **loop_context,
                      netsnmp_variable_list *index,
                      netsnmp_iterator_info *data)
 {
-    int i = (int)*loop_context;
+    int i = (intptr_t)*loop_context;
     netsnmp_variable_list *idx;
     long port;
 
@@ -419,7 +419,7 @@ tcpTable_next_entry( void **loop_context,
      * and update the loop context ready for the next one.
      */
     *data_context = (void*)&tcp_head[i];
-    *loop_context = (void*)++i;
+    *loop_context = (void*)(intptr_t)++i;
 
     return index;
 }
